Make timeShift appear in dashboard layouts
